+++ package.json
@@ -91,6 +91,7 @@
     "show-errors": "npm test 1> /dev/null",
     "test": "npm run format && npm run lint && npm run unittest",
     "traverse": "node --loader=ts-node/esm --no-warnings=ExperimentalWarning scripts/traverse.ts",
-    "update": "node --loader=ts-node/esm --no-warnings=ExperimentalWarning scripts/update.ts"
+    "update": "node --loader=ts-node/esm --no-warnings=ExperimentalWarning scripts/update.ts",
+    "update-browser-releases": "node --loader=ts-node/esm --no-warnings=ExperimentalWarning scripts/update-browser-releases/index.ts"
   }
 }
